remove complex parsing method, add missing column to default requests list, split requests and default requests in 2 tabs on the dedicated frontend page
